Fix handling zoneinfo files generated by zic 2014c and later.
Fix an error in the conversion of -2**63 from an unsigned integer.

Ignore timestamps that are either far in the past or far in the future
and which currently cause problems for the transition index. This is
primarily intended to ignore the far in the past transition added in zic
2014c (at timestamp -2**63 in zic 2014c and at approximate time of the
big bang from zic 2014d).