Remove TimezonePeriod range comparison methods.
These methods used to be used internally in old versions of TZInfo to
identify the appropriate TimezonePeriod for a given time.

They are unlikely to have been used extensively outside of TZInfo, since
it is only possible to obtain TimezonePeriod instances by specifying a
local or UTC time for which they would be valid.