Extract the DSL used by TZInfo::Data from TransitionDataTimezoneInfo.
An instance of TimezoneDefinerFormat1 is now yielded by
TimezoneDefinition::ClassMethods.timezone instead of a
TransitionDataTimezoneInfo instance.

TransitionDataTimezoneInfo.new now takes an Array of TimezoneTransitions
or a TimezoneOffset.

ZoneinfoTimezoneInfo has been changed to use the new
TransitionDataTimezoneInfo constructor.

The tests for ZoneinfoTimezoneInfo have been simplified so that instead
of checking the results of calculations, they check that the correct
transitions or constant offset have been passed to the
TransitionDataTimezoneInfo constructor.