Move data source related classes into a new DataSources sub-module.
The *Info classes are used to return data from DataSource
implementations. They are not returned as part of the primary interface
and should only be of interest to those creating custom DataSources.

Other than setting zoneinfo search paths, the RubyDataSource and
ZoneinfoDataSource classes are not referenced directly.

ZoneinfoReader is only used internally.