Fix SecurityErrors when loading data in safe mode.
Recent Ruby releases will raise a "SecurityError: Insecure operation -
require" exception when requiring a relative path name in safe mode.

Calculate the full path to the data file for use in require.

Similar to the changes made for 2.0.0, but doesn't raise an exception if
tzinfo/data can't be found when RubyDataSource is initialized.