Support "slim" zoneinfo files produced by default by zic >= 2020b.
Use the POSIX-style TZ string to calculate transitions after the last
defined transition contained within the file. At the moment these
transitions are generated when the file is loaded. In a later release
this will likely be changed to calculate transitions on demand.

Use the 64-bit section of zoneinfo files regardless of whether the
runtime supports 64-bit Times. The 32-bit section is empty in "slim"
zoneinfo files.

Resolves #120.