Eliminate Object#untaint deprecation warnings on JRuby 9.4.0.0.
Despite using the UntaintExt refinement that replaces Object#untaint,
JRuby 9.4.0.0 calls the original Object#untaint method and outputs
warnings as a result:

warning: Object#untaint is deprecated and will be removed in Ruby 3.2

This is caused by the unorthodox `send(:using, UntaintExt)` approach
being used to deal with issues with older JRuby versions (see #114).
JRuby 9.4.0.0 doesn't detect this as using the refinement. This will be
allowed again in JRuby 9.4.1.0 (see jruby/jruby#7599).

Replace the UntaintExt refinement with an untaint method in
RubyCoreSupport (now reinstated having previously been removed in
2.0.0).

Change tests for handling of tainted inputs to skip when taint/untaint
is undefined or a no-op. There's no point in running these tests unless
the inputs are actually tainted. Remove the (test-only) TaintExt
refinement too.

Resolves #145.